Move delete records to delete_record.py
